View 2 Done

+++ backend/src/main/java/synaos/calendar/data/CelebrationsRepository.java
@@ -3,5 +3,5 @@ package synaos.calendar.data;
 import org.springframework.data.jpa.repository.JpaRepository;
 import synaos.entities.Celebrations;
 
-public interface CelebrationsRepository extends JpaRepository<Celebrations, Long> {
+public interface CelebrationsRepository extends JpaRepository<Celebrations, Integer> {
 }
